% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DELTD.R
\docType{package}
\name{DELTD-package}
\alias{DELTD}
\alias{DELTD-package}
\title{DELTD}
\description{
A collection of asymmetrical kernels belong to lifetime distributions for kernel density estimation is presented. i.e. \code{\link{plot.BS}}, \code{\link{plot.Erlang}},
\code{\link{plot.Gamma}}, \code{\link{plot.LN}} and for plotting all densities at same time \code{\link{dencomb}}. Estimated values can also observed by using
\code{\link{BS}}, \code{\link{Gamma}}, \code{\link{Erlang}} and \code{\link{LN}}. For calculating mean squared error by using different kernels functions are \code{\link{mseBS}}, \code{\link{mseEr}},
\code{\link{mseGamma}} and \code{\link{mseLN}}.
}
\details{
Kernel Density Estimation using Lifetime Distributions
}
\references{
\itemize{
\item Jin, X.; Kawczak, J. 2003. Birnbaum-Saunders & Lognormal kernel estimators for modeling durations in high frequency financial data. \emph{Annals of Economics and Finance} \strong{4}, 103–124.
\item Salha, R. B.; Ahmed, E. S.; Alhoubi, I. M. 2014. Hazard rate function estimation ksing Erlang Kernel. \emph{Pure Mathematical Sciences} \strong{3} (4), 141–152.
\item Chen, S. X. 2000. Probability density function estimation using Gamma kernels. \emph{Annals of the Institute of Statistical Mathematics} \strong{52} (3), 471-480.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=DELTD}
}

}
\author{
Javaria Ahmad Khan, Atif Akbar.
}
