% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DELTD.R
\name{denBS}
\alias{denBS}
\title{Plot Density by Birnbaum-Saunders kernel.}
\usage{
denBS(y, k, h)
}
\arguments{
\item{y}{a numeric vector of positive values.}

\item{k}{gird points.}

\item{h}{the bandwidth}
}
\value{
Plot the density by using Birnbaum-Saunders kernel
}
\description{
Plot Kernel density by using Birnbaum-Saunders Kernel.
}
\examples{
y<-rexp(23,1)
h<-0.79 * IQR(y) * length(y) ^ (-1/5)
denBS(y,80,h)
}
