% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DELTD.R
\name{dencomb}
\alias{dencomb}
\title{Plot the densities for comparison.}
\usage{
dencomb(y, k, h, comb)
}
\arguments{
\item{y}{a numeric vector of positive values.}

\item{k}{gird points.}

\item{h}{the bandwidth}

\item{comb}{mention the combination which kernel estimated densities are to be compared. If Lognormal and Birnbaum-Saunders kernel densities
are to be compared along with real density then use "TLB". If Lognormal and Erlang then use "TLE". If Lognormal and Gamma to be compared
then use "TLG". For Birnbaum-Saunders and Erlang use "TBE". For Birnbaum-Saunders and Gamma then use "TBG". For Erlang and Gamma use "TEG".
For Lognormal, Birnbaum-Saunders and Erlang use "TLBE". For Lognormal, Birnbaum-Saunders and Gamma use "TLBG". For Birnbaum-Saunders, Erlang and Gamma
use "TBEG". To compare all densities in one graph use "TLBEG".}
}
\value{
Plot of Estimated Densities with Real Data Density.
}
\description{
Plot the Estimated Densities with Real Density for Comparison .
}
\examples{
\dontshow{
y<-rexp(10,1)
h<-0.79 * IQR(y) * length(y) ^ (-1/5)
dencomb(y,20,h,"TLB")}
}
