% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DELTD.R
\name{denEr}
\alias{denEr}
\title{Plot Density by Erlang kernel.}
\usage{
denEr(y, k, h)
}
\arguments{
\item{y}{a numeric vector of positive values.}

\item{k}{gird points.}

\item{h}{the bandwidth}
}
\value{
Plot the density by using Erlang kernel
}
\description{
Plot Kernel density by using Erlang Kernel.
}
\examples{
y<-rexp(23,1)
h<-0.79 * IQR(y) * length(y) ^ (-1/5)
denEr(y,80,h)
}
