% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DELTD.R
\name{denGamma}
\alias{denGamma}
\title{Density Plot by Gamma kernel.}
\usage{
denGamma(y, k, h)
}
\arguments{
\item{y}{a numeric vector of positive values.}

\item{k}{gird points.}

\item{h}{the bandwidth}
}
\value{
Plot the density by using Gamma kernel
}
\description{
Plot Kernel density by using Gamma Kernel.
}
\examples{
y<-rexp(23,1)
h<-0.79 * IQR(y) * length(y) ^ (-1/5)
denGamma(y,80,h)
}
