% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DELTD.R
\name{denLN}
\alias{denLN}
\title{Plot Density by Lognormal kernel.}
\usage{
denLN(y, k, h)
}
\arguments{
\item{y}{a numeric vector of positive values.}

\item{k}{gird points.}

\item{h}{the bandwidth}
}
\value{
Plot the density by using Lognormal kernel
}
\description{
Plot Kernel density by using Lognormal Kernel.
}
\examples{
y<-rexp(23,1)
h<-0.79 * IQR(y) * length(y) ^ (-1/5)
denLN(y,80,h)
}
