\name{suicide}

\docType{data}
\title{Data of control Patients of Suicide Study}
\description{A dataset contains a lengths of treatment spells (in days) of control patients of suicide study.}

\source{Silverman, B. W. (1986). Density Estimation. Chapman and Hall.}

\examples{data(suicide)}
\keyword{datasets}
