% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DELTD.R
\docType{data}
\name{TUNA}
\alias{TUNA}
\title{Data of Tuna fish}
\format{
A vector with 64 observations
}
\usage{
TUNA
}
\description{
Data is about Tuna, which is saltwater fish. Its seasonal migration is between waters off the coast of Australia and the Indian Ocean. The data represents a line transect aerial survey of
Southern Bluefin Tuna in the Great Australian Bight in summer when the tuna tend to stay on the surface. The abundance D is measured by \deqn{D = \frac{N}{A}} , where N is the total number of surface
schools in the Bight and A is the survey area. To estimate D, an aircraft with two spotters on board is used to fly randomly allocated transect lines to detect tuna schools.
Each school sighted from transect is counted and its perpendicular distance to transect is measured.
}
\references{
Buckland, S. T.; Burnham, K. P.; Anderson, D. R.; Laake, J. L. 1993. \emph{Density Estimation using Distance Sampling}. Chapman & Hall, London.
}
\keyword{datasets}
